Refactors package structure for consent and financial data response classes
