+++ accountservice/src/main/java/com/micronauticals/accountservice/service/SetuServiceImpl/SetuAuthServiceImpl.java
@@ -1,8 +1,19 @@
 package com.micronauticals.accountservice.service.SetuServiceImpl;
 import com.micronauticals.accountservice.Dto.request.ConsentRequestDTO;
 import com.micronauticals.accountservice.Dto.request.SetuLoginRequest;
-import com.micronauticals.accountservice.Dto.response.*;
+import com.micronauticals.accountservice.Dto.response.consent.ConsentDataSessionResponseDTO;
+import com.micronauticals.accountservice.Dto.response.consent.ConsentResponse;
+import com.micronauticals.accountservice.Dto.response.consent.ConsentStatusResponseDTO;
+import com.micronauticals.accountservice.Dto.response.consent.RevokeConsentResponse;
+import com.micronauticals.accountservice.Dto.response.financialdata.FIPResponseDTO;
+import com.micronauticals.accountservice.Dto.response.financialdata.SetuLoginResponse;
+import com.micronauticals.accountservice.entity.consent.Consent;
+import com.micronauticals.accountservice.entity.financialdata.FiDataBundle;
 import com.micronauticals.accountservice.exception.SetuLoginException;
+import com.micronauticals.accountservice.mapper.ConsentDtoToEntity;
+import com.micronauticals.accountservice.mapper.FIPResponseDtoToEntityMapper;
+import com.micronauticals.accountservice.repository.ConsentRepository;
+import com.micronauticals.accountservice.repository.FIDataRepository;
 import com.micronauticals.accountservice.service.SetuServiceInterface.SetuAuthService;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
@@ -10,8 +21,8 @@ import org.slf4j.LoggerFactory;
 import org.springframework.http.*;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
-import org.springframework.web.servlet.View;
 import reactor.core.publisher.Mono;
+import reactor.core.scheduler.Schedulers;
 
 import java.time.Duration;
 import java.util.Map;
@@ -20,11 +31,14 @@ import java.util.Map;
 public class SetuAuthServiceImpl implements SetuAuthService {
 
     private final WebClient.Builder webClientBuilder;
+    private final ConsentRepository consentRepository;
+    private final ConsentDtoToEntity consentDtoToEntity;
+    private final FIDataRepository fiDataRepository;
+    private final FIPResponseDtoToEntityMapper fipResponseDtoToEntityMapper;
 
     private static final String SETU_LOGIN_URL = "https://orgservice-prod.setu.co/v1/users/login";
     private static final String SETU_CONSENT_URL = "https://fiu-sandbox.setu.co/v2/consents";
     private static final Logger log = LoggerFactory.getLogger(SetuAuthServiceImpl.class);
-    private final View error;
 
     private String accessToken;
     private String refreshToken;
@@ -63,7 +77,7 @@ public class SetuAuthServiceImpl implements SetuAuthService {
                 throw new SetuLoginException("Missing access token in Setu response");
             }
 
-            this.accessToken = (String) result.get("access_token"); // ✅ store token
+            this.accessToken = (String) result.get("access_token");
             this.refreshToken = (String) result.get("refresh_token");
 
             log.info("Received access token from Setu");
@@ -101,6 +115,12 @@ public class SetuAuthServiceImpl implements SetuAuthService {
                             });
                 })
                 .bodyToMono(ConsentResponse.class)
+                .flatMap(response -> Mono.fromCallable(() -> {
+                    Consent consent = consentDtoToEntity.mapToEntity(response);
+                    consentRepository.save(consent);
+                    log.info("Consent saved in DB with ID: {}", consent.getId());
+                    return response;
+                }).subscribeOn(Schedulers.boundedElastic()))
                 .doOnNext(response -> log.info("Consent created successfully: {}", response))
                 .doOnError(error -> log.error("Error occurred during consent creation", error));
 
@@ -165,7 +185,7 @@ public class SetuAuthServiceImpl implements SetuAuthService {
     }
 
     @Override
-    public Mono<FinancialDataFetchResponseDTO> getFiData(String sessionId){
+    public Mono<FIPResponseDTO> getFiData(String sessionId){
 
         if (accessToken == null) {
             return Mono.error(new SetuLoginException("Access token not available. Please login first."));
@@ -188,7 +208,13 @@ public class SetuAuthServiceImpl implements SetuAuthService {
                                 return Mono.error(new RuntimeException("Error fetching consent status: " + errorBody));
                             });
                 })
-                .bodyToMono(FinancialDataFetchResponseDTO.class)
+                .bodyToMono(FIPResponseDTO.class)
+                .flatMap(response -> Mono.fromCallable(() -> {
+                    FiDataBundle fiDataBundle = fipResponseDtoToEntityMapper.mapToEntity(response); // Ensure this returns FiDataBundle
+                    fiDataRepository.save(fiDataBundle); // Save the correct type
+                    log.info("Data saved in DB with ID: {}", fiDataBundle.getId());
+                    return response; // Return the original FIPResponseDTO
+                }).subscribeOn(Schedulers.boundedElastic()))
                 .doOnNext(response -> log.info("Consent status fetched successfully: {}", response))
                 .doOnError(error -> log.error("Error occurred while fetching consent status", error));
